+++ packages/applications/ssr/src/components/pages/réseau/raccordement/transmettre/transmettreDateMiseEnService/TransmettreDateMiseEnService.form.tsx
@@ -1,0 +1,65 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import { FC } from 'react';
+import { Button } from '@codegouvfr/react-dsfr/Button';
+
+import { Routes } from '@potentiel-applications/routes';
+import { now } from '@potentiel-libraries/iso8601-datetime';
+
+import { Form } from '@/components/atoms/form/Form';
+import { InputDate } from '@/components/atoms/form/InputDate';
+import { SubmitButton } from '@/components/atoms/form/SubmitButton';
+
+import { transmettreDateMiseEnServiceAction } from './transmettreDateMiseEnService.action';
+import { TransmettreDateMiseEnServicePageProps } from './TransmettreDateMiseEnService.page';
+
+type TransmettreDateMiseEnServiceFormProps = {
+  projet: TransmettreDateMiseEnServicePageProps['projet'];
+  dossierRaccordement: TransmettreDateMiseEnServicePageProps['dossierRaccordement'];
+};
+
+export const TransmettreDateMiseEnServiceForm: FC<TransmettreDateMiseEnServiceFormProps> = ({
+  projet: { identifiantProjet, dateDésignation },
+  dossierRaccordement: { référence, miseEnService },
+}) => {
+  const router = useRouter();
+  return (
+    <Form
+      method="POST"
+      heading="Transmettre la date de mise en service"
+      action={transmettreDateMiseEnServiceAction}
+      onSuccess={() => router.push(Routes.Raccordement.détail(identifiantProjet))}
+    >
+      <input type="hidden" name="identifiantProjet" value={identifiantProjet} />
+      <input type="hidden" name="referenceDossier" value={référence} />
+      <input type="hidden" name="dateDesignation" value={dateDésignation} />
+
+      <InputDate
+        label="Date de mise en service"
+        nativeInputProps={{
+          type: 'date',
+          name: 'dateMiseEnService',
+          defaultValue: miseEnService,
+          min: dateDésignation,
+          max: now(),
+          required: true,
+          'aria-required': true,
+        }}
+      />
+
+      <div className="flex flex-col md:flex-row gap-4 md:mt-4">
+        <Button
+          priority="secondary"
+          linkProps={{
+            href: Routes.Raccordement.détail(identifiantProjet),
+          }}
+          iconId="fr-icon-arrow-left-line"
+        >
+          Retour aux dossiers de raccordement
+        </Button>
+        <SubmitButton>Transmettre</SubmitButton>
+      </div>
+    </Form>
+  );
+};
