+++ packages/applications/ssr/src/components/pages/réseau/raccordement/transmettre/transmettreDateMiseEnService/TransmettreDateMiseEnService.page.tsx
@@ -1,24 +1,16 @@
-'use client';
-
 import Alert from '@codegouvfr/react-dsfr/Alert';
-import Button from '@codegouvfr/react-dsfr/Button';
-import { useRouter } from 'next/navigation';
 
-import { Routes } from '@potentiel-applications/routes';
-import { Iso8601DateTime, now } from '@potentiel-libraries/iso8601-datetime';
+import { Iso8601DateTime } from '@potentiel-libraries/iso8601-datetime';
 
 import { FormattedDate } from '@/components/atoms/FormattedDate';
-import { Form } from '@/components/atoms/form/Form';
-import { InputDate } from '@/components/atoms/form/InputDate';
-import { SubmitButton } from '@/components/atoms/form/SubmitButton';
 import { ProjetBanner } from '@/components/molecules/projet/ProjetBanner';
 import { ColumnPageTemplate } from '@/components/templates/ColumnPage.template';
 
 import { TitrePageRaccordement } from '../../TitrePageRaccordement';
 
-import { transmettreDateMiseEnServiceAction } from './transmettreDateMiseEnService.action';
+import { TransmettreDateMiseEnServiceForm } from './TransmettreDateMiseEnService.form';
 
-export type TransmettreDateMiseEnServiceProps = {
+export type TransmettreDateMiseEnServicePageProps = {
   projet: {
     identifiantProjet: string;
     dateDésignation: Iso8601DateTime;
@@ -31,102 +23,66 @@ export type TransmettreDateMiseEnServiceProps = {
 };
 
 export const TransmettreDateMiseEnServicePage = ({
-  projet: { dateDésignation, identifiantProjet },
-  dossierRaccordement: { référence, miseEnService },
+  projet,
+  dossierRaccordement,
   intervalleDatesMeSDélaiCDC2022,
-}: TransmettreDateMiseEnServiceProps) => {
-  const router = useRouter();
-
-  return (
-    <ColumnPageTemplate
-      banner={<ProjetBanner identifiantProjet={identifiantProjet} />}
-      heading={<TitrePageRaccordement />}
-      leftColumn={{
-        children: (
-          <Form
-            method="POST"
-            heading="Transmettre la date de mise en service"
-            action={transmettreDateMiseEnServiceAction}
-            onSuccess={() => router.push(Routes.Raccordement.détail(identifiantProjet))}
-          >
-            <input type="hidden" name="identifiantProjet" value={identifiantProjet} />
-            <input type="hidden" name="referenceDossier" value={référence} />
-            <input type="hidden" name="dateDesignation" value={dateDésignation} />
-
-            <InputDate
-              label="Date de mise en service"
-              nativeInputProps={{
-                type: 'date',
-                name: 'dateMiseEnService',
-                defaultValue: miseEnService,
-                min: dateDésignation,
-                max: now(),
-                required: true,
-                'aria-required': true,
-              }}
-            />
-
-            <div className="flex flex-col md:flex-row gap-4 md:mt-4">
-              <Button
-                priority="secondary"
-                linkProps={{
-                  href: Routes.Raccordement.détail(identifiantProjet),
-                }}
-                iconId="fr-icon-arrow-left-line"
-              >
-                Retour aux dossiers de raccordement
-              </Button>
-              <SubmitButton>Transmettre</SubmitButton>
-            </div>
-          </Form>
-        ),
-      }}
-      rightColumn={{
-        children: (
-          <Alert
-            severity="info"
-            small
-            description={
-              <div className="py-4 text-justify">
-                <ul className="flex flex-col gap-3">
-                  {intervalleDatesMeSDélaiCDC2022 && (
-                    <li>
-                      Si le projet{' '}
-                      <span className="font-bold">
-                        a bénéficié du délai supplémentaire relatif du cahier des charges du
-                        30/08/2022
-                      </span>
-                      , la saisie d'une date de mise en service non comprise entre le{' '}
-                      <FormattedDate
-                        className="font-bold"
-                        date={intervalleDatesMeSDélaiCDC2022.min}
-                      />{' '}
-                      et le{' '}
-                      <FormattedDate
-                        className="font-bold"
-                        date={intervalleDatesMeSDélaiCDC2022.max}
-                      />{' '}
-                      peut remettre en cause l'application de ce délai et entraîner une modification
-                      de la date d'achèvement du projet.
-                    </li>
-                  )}
+}: TransmettreDateMiseEnServicePageProps) => (
+  <ColumnPageTemplate
+    banner={<ProjetBanner identifiantProjet={projet.identifiantProjet} />}
+    heading={<TitrePageRaccordement />}
+    leftColumn={{
+      children: (
+        <TransmettreDateMiseEnServiceForm
+          projet={projet}
+          dossierRaccordement={dossierRaccordement}
+        />
+      ),
+    }}
+    rightColumn={{
+      children: (
+        <Alert
+          severity="info"
+          small
+          description={
+            <div className="py-4 text-justify">
+              <ul className="flex flex-col gap-3">
+                {intervalleDatesMeSDélaiCDC2022 && (
                   <li>
                     Si le projet{' '}
                     <span className="font-bold">
-                      n'a pas bénéficié du délai supplémentaire relatif du cahier des charges du
+                      a bénéficié du délai supplémentaire relatif du cahier des charges du
                       30/08/2022
                     </span>
-                    , la saisie d'une date de mise en service doit être comprise entre la date de
-                    désignation du projet{' '}
-                    <FormattedDate className="font-bold" date={dateDésignation} /> et{' '}
-                    <span className="font-bold">ce jour</span>.
+                    , la saisie d'une date de mise en service non comprise entre le{' '}
+                    <FormattedDate
+                      className="font-bold"
+                      date={intervalleDatesMeSDélaiCDC2022.min}
+                    />{' '}
+                    et le{' '}
+                    <FormattedDate
+                      className="font-bold"
+                      date={intervalleDatesMeSDélaiCDC2022.max}
+                    />{' '}
+                    peut remettre en cause l'application de ce délai et entraîner une modification
+                    de la date d'achèvement du projet.
                   </li>
-                </ul>
-              </div>
-            }
-          />
-        ),
-      }}
-    />
-  );
-};
+                )}
+                <li>
+                  Si le projet{' '}
+                  <span className="font-bold">
+                    n'a pas bénéficié du délai supplémentaire relatif du cahier des charges du
+                    30/08/2022
+                  </span>
+                  , la saisie d'une date de mise en service doit être comprise entre la date de
+                  désignation du projet{' '}
+                  <FormattedDate className="font-bold" date={projet.dateDésignation} /> et{' '}
+                  <span className="font-bold">ce jour</span>.
+                </li>
+              </ul>
+            </div>
+          }
+        />
+      ),
+    }}
+  />
+);
